+++ be/src/vec/sink/vdata_stream_sender.cpp
@@ -230,7 +230,7 @@ Status Channel::close(RuntimeState* state) {
 BlockSerializer::BlockSerializer(pipeline::ExchangeSinkLocalState* parent, bool is_local)
         : _parent(parent), _is_local(is_local), _batch_size(parent->state()->batch_size()) {}
 
-Status BlockSerializer::next_serialized_block(Block* block, PBlock* dest, int num_receivers,
+Status BlockSerializer::next_serialized_block(Block* block, PBlock* dest, size_t num_receivers,
                                               bool* serialized, bool eos,
                                               const std::vector<uint32_t>* rows) {
     if (_mutable_block == nullptr) {
@@ -261,7 +261,7 @@ Status BlockSerializer::next_serialized_block(Block* block, PBlock* dest, int nu
     return Status::OK();
 }
 
-Status BlockSerializer::serialize_block(PBlock* dest, int num_receivers) {
+Status BlockSerializer::serialize_block(PBlock* dest, size_t num_receivers) {
     if (_mutable_block && _mutable_block->rows() > 0) {
         auto block = _mutable_block->to_block();
         RETURN_IF_ERROR(serialize_block(&block, dest, num_receivers));
@@ -272,7 +272,7 @@ Status BlockSerializer::serialize_block(PBlock* dest, int num_receivers) {
     return Status::OK();
 }
 
-Status BlockSerializer::serialize_block(const Block* src, PBlock* dest, int num_receivers) {
+Status BlockSerializer::serialize_block(const Block* src, PBlock* dest, size_t num_receivers) {
     SCOPED_TIMER(_parent->_serialize_batch_timer);
     dest->Clear();
     size_t uncompressed_bytes = 0, compressed_bytes = 0;
